Converted some of the existing tests into Boost tests
