+++ Common/test/MPI_global_fixture.hpp
@@ -1,8 +1,8 @@
 /*!
- * \file unit_test_config.cpp
- * \brief Test CConfig ctor that gives valid config object without input file.
- * \author T. A. Oliver
- * \version 5.0.0 "Raven"
+ * \file MPI_global_fixture.hpp
+ * \brief This header defines a global fixture that allows use of MPI
+ * \author C. Pederson
+ * \version 4.3.0 "Cardinal"
  *
  * SU2 Lead Developers: Dr. Francisco Palacios (Francisco.D.Palacios@boeing.com).
  *                      Dr. Thomas D. Economon (economon@stanford.edu).
@@ -31,23 +31,52 @@
  * License along with SU2. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#define BOOST_TEST_MODULE Unit_test_config
-#include "MPI_global_fixture.hpp"
-
-#include "../../Common/include/config_structure.hpp"
-
 #ifdef BUILD_TESTS
+#include "boost/test/included/unit_test.hpp"
 
-BOOST_GLOBAL_FIXTURE( MPIGlobalFixture );
+#ifdef HAVE_MPI
+#include <mpi.h>
+#endif
 
-BOOST_AUTO_TEST_CASE(Unit_Test_Config) {
+/*
+ * After including this header, include the line:
+ *
+ * BOOST_GLOBAL_FIXTURE( MPIGlobalFixture );
+ *
+ * to use this fixture in addition to any other fixtures.  This macro will
+ * have file scope, affecting all tests and test suites in the current file.
+ */
 
-  CConfig* test_config;
-  test_config = new CConfig();
+// A fixture to automate setup/teardown of MPI.
+struct MPIGlobalFixture {
 
-  BOOST_CHECK(test_config->GetKind_Regime() == COMPRESSIBLE);
+  /**
+   * Setup MPI
+   *
+   * When this fixture is used, MPI will be setup whenever a test starts.
+   */
+  MPIGlobalFixture() {
+#ifdef HAVE_MPI
+  MPI_Init(NULL,NULL);
+  BOOST_TEST_MESSAGE("Calling MPI_Init...");
+#endif
+  }
 
-  delete test_config;
-}
+  /**
+   * Tear down MPI
+   *
+   * When this fixture is used, MPI will be finalize whenever a test starts.
+   * This will happen whether or not a particular test fails or throws an
+   * exception.  If the test is expected to call MPI_Abort or MPI_Finalize
+   * itself, you probably shouldn't use this fixture.
+   */
+  ~MPIGlobalFixture() {
+#ifdef HAVE_MPI
+  MPI_Finalize();
+  BOOST_TEST_MESSAGE("Calling MPI_Finalize...");
+#endif
+  }
+
+};
 
 #endif
