+++ Makefile.am
@@ -35,7 +35,6 @@
 
 AUTOMAKE_OPTIONS = subdir-objects
 ACLOCAL_AMFLAGS = ${ACLOCAL_FLAGS} -I m4
-EXTRA_DIST = bootstrap
 
 # build third-party optional dependencies first
 SUBDIRS = externals
