+++ configure.ac
@@ -140,6 +140,9 @@ AC_ARG_ENABLE(normal,
 AC_ARG_ENABLE(testing,
     AS_HELP_STRING([--enable-testing], [enable testing with the Boost UTF (default = no)]),
     [build_TESTS="yes"], [build_TESTS="no"])
+AC_ARG_WITH(Boost-include,
+    AS_HELP_STRING([--with-Boost-include[=ARG]], [Boost include directory, ARG = path to ../boost/include]),
+    [with_Boost_include=$withval], [with_Boost_include="no"])
 
 
 # Check for the old  MPI option so that we can throw an error
@@ -427,6 +430,35 @@ AC_SUBST([JSONCPP_LD])
 
 ##########################
 
+AM_CONDITIONAL([BUILD_TESTS], [test $build_TESTS != "no"])
+
+# check Boost include folder for correctness
+Boostheader=boost/test/included/unit_test_framework.hpp
+have_Boost="no"
+if test "$build_TESTS" != "no"
+then
+    if test "$with_Boost_include" != "no"
+    then
+        AC_CHECK_FILE([$with_Boost_include/$Boostheader],[have_Boost='yes'],[have_Boost='no'])
+        if test "$have_Boost" == "no"
+        then
+            AC_MSG_ERROR([Boost include folder specified but header file not found.])
+        fi
+    else
+        have_Boost="no"
+    fi
+    
+    if test "$have_Boost" != "no"
+    then
+        BOOST_CXX="-I"$with_Boost_include""
+    else
+        BOOST_CXX=
+    fi
+fi
+AC_SUBST([BOOST_CXX])
+
+##########################
+
 # check for lib
 #LAPACKlib=liblapacke.a
 LAPACKlib=libmkl_lapack95_lp64.a
@@ -530,7 +562,6 @@ AM_CONDITIONAL([BUILD_SOL],[test $build_SOL != "no"])
 AM_CONDITIONAL([BUILD_GEO],[test $build_GEO != "no"])
 AM_CONDITIONAL([BUILD_PY],[test $build_PY != "no"])
 AM_CONDITIONAL([BUILD_PY_WRAPPER], [test $build_PY_WRAPPER != "no"])
-AM_CONDITIONAL([BUILD_TESTS], [test $build_TESTS != "no"])
 
 ###########################
